// src/models/alteryx-tool-type.ts
export enum AlteryxToolType {
  AppendFields = 'AppendFields',
  CrossTab = 'CrossTab',
  DbFileInput = 'DbFileInput',
  DbFileOutput = 'DbFileOutput',
  DynamicInput = 'DynamicInput',
  Filter = 'Filter',
  FindReplace = 'FindReplace',
  Formula = 'Formula',
  GenerateRows = 'GenerateRows',
  Join = 'Join',
  MultiFieldFormula = 'MultiFieldFormula',
  MultiRowFormula = 'MultiRowFormula',
  RandomRecords = 'RandomRecords',
  RunningTotal = 'RunningTotal',
  Sample = 'Sample',
  Select = 'Select',
  Sort = 'Sort',
  Summarize = 'Summarize',
  Transpose = 'Transpose',
  Union = 'Union',
  Unique = 'Unique',
  Unknown = 'Unknown',
}